include thirdPartyDataDeclarationStatus field in the creative service
