+++ lib/client/common/types/general.type.ts
@@ -22,31 +22,6 @@ import {
 } from "../enums";
 
 /**
- * The status of the ThirdPartyDataDeclaration associated with a given creative.
- *
- * This is calculated by comparing the companies detected via automated scanning/parsing, with the companies declared in the ThirdPartyDataDeclaration.
- */
-export enum ThirdPartyDataDeclarationStatus {
-  /**
-   * The value returned if the actual value is not exposed by the requested API version. 
-   */
-  UNKNOWN = "UNKNOWN",
-  /**
-   * At least one detected ad technology provider wasn’t declared at the network level or on the creative's ThirdPartyDataDeclaration
-   */
-  INCOMPLETE = "INCOMPLETE",
-  /**
-   * All detected ad technology providers have been declared, either at the network level or on the creative's ThirdPartyDataDeclaration
-   */
-  COMPLETE = "COMPLETE",
-  /**
-   * This entity has not been recently scanned.
-   * There is either no scanning data for this creative, or the scanning data is stale. This can happen if the creative is not associated with any line items that are active or eligible to serve.
-   */
-  UNSCANNED = "UNSCANNED",
-}
-
-/**
  * Represents a set of declarations about what (if any) third party companies are associated with a given creative.
  *
  * This can be set at the network level, as a default for all creatives, or overridden for a particular creative.
